template分割、画像オブジェクト生成後にCanvasをレンダリングするよう修正 #19
