+++ src/Puzzle.js
@@ -1,219 +1,10 @@
-import React, { useState, useEffect } from 'react';
-import { Divider } from '@material-ui/core';
-import Konva from 'konva';
-import useImage from 'use-image';
-import _shuffle from 'lodash/shuffle';
-import PuzzleGuide from './organisms/PuzzleGuide';
-import SelectDifficultyModal from './organisms/SelectDifficultyModal';
-import CompleteModal from './organisms/CompleteModal';
-import PauseModal from './organisms/PauseModal';
-import PuzzleCanvas from './organisms/PuzzleCanvas';
+import React from 'react';
+import PuzzleTemplate from './templates/PuzzleTemplate';
 
 const Puzzle = () => {
-  const [hour, setHour] = useState('00');
-  const [minutes, setMinutes] = useState('00');
-  const [seconds, setSeconds] = useState('00');
-  const [timerId, setTimerId] = useState(0);
-  let time = 0;
-  const [backupTime, setBackupTime] = useState(0);
+  const imageUrl = `${process.env.PUBLIC_URL}/photo0000_6339.png`;
 
-  const parseHours = () => {
-    return `00${parseInt(time / 60 / 60, 10)}`.slice(-2);
-  };
-  const parseMinutes = () => {
-    return `00${parseInt((time / 60) % 60, 10)}`.slice(-2);
-  };
-  const parseSeconds = () => {
-    return `00${parseInt(time % 60, 10)}`.slice(-2);
-  };
-  const timeCount = () => {
-    if (!time && backupTime) {
-      time = backupTime;
-    }
-    time += 1;
-    setHour(parseHours());
-    setMinutes(parseMinutes());
-    setSeconds(parseSeconds());
-    setBackupTime(time);
-  };
-  const timerStart = () => {
-    setTimerId(setInterval(timeCount, 1000));
-  };
-  const timeStop = () => {
-    clearInterval(timerId);
-  };
-
-  const [difficultyModalOpen, setDifficultyModalOpen] = useState(true);
-  const [pauseModalOpen, setPauseModalOpen] = useState(false);
-  const [completeModalOpen, setCompleteModalOpen] = useState(false);
-  const [matchPieceCount, setMatchPieceCount] = useState(0);
-  const [pieceXCount, setPieceXCount] = useState(0);
-  const [pieceYCount, setPieceYCount] = useState(0);
-  const [pieceSize, setPieceSize] = useState(0);
-  const [shuffledPieceInfo, setShuffledPieceInfo] = useState([]);
-
-  const handlePauseAction = () => {
-    timeStop();
-    setPauseModalOpen(true);
-  };
-  useEffect(() => {
-    const initialPiece = () => {
-      const pieceInfo = [];
-      const randomKey = Math.random().toString(32).substring(2);
-      for (let y = 0; y < pieceYCount; y += 1) {
-        for (let x = 0; x < pieceXCount; x += 1) {
-          pieceInfo.push({
-            id: `${y}-${x}-${randomKey}`,
-            crop: {
-              x: pieceSize * x,
-              y: pieceSize * y,
-              width: pieceSize,
-              height: pieceSize,
-            },
-            width: pieceSize,
-            height: pieceSize,
-          });
-        }
-      }
-      setShuffledPieceInfo(_shuffle(pieceInfo));
-    };
-    if (pieceXCount && pieceYCount && pieceSize) {
-      initialPiece();
-    }
-  }, [pieceXCount, pieceYCount, pieceSize]);
-
-  useEffect(() => {
-    const complete = () => {
-      setCompleteModalOpen(true);
-    };
-    if (matchPieceCount && matchPieceCount === pieceXCount * pieceYCount) {
-      timeStop();
-      setTimeout(complete, 1500);
-    }
-  }, [matchPieceCount]);
-
-  const handleSelectDifficultyAction = (diffculty) => {
-    switch (diffculty) {
-      case 'easy':
-        setPieceXCount(6);
-        setPieceYCount(4);
-        setPieceSize(120);
-        break;
-      case 'normal':
-        setPieceXCount(9);
-        setPieceYCount(6);
-        setPieceSize(80);
-        break;
-      case 'hard':
-        setPieceXCount(12);
-        setPieceYCount(8);
-        setPieceSize(60);
-        break;
-      default:
-        console.log('error');
-    }
-    // initialPiece();
-    setDifficultyModalOpen(false);
-    timerStart();
-  };
-  const handlePauseReleseAction = () => {
-    setPauseModalOpen(false);
-    timerStart();
-  };
-  const handleRestartAction = () => {
-    setCompleteModalOpen(false);
-    setHour('00');
-    setMinutes('00');
-    setSeconds('00');
-    setBackupTime(0);
-    setPieceXCount(0);
-    setPieceYCount(0);
-    setMatchPieceCount(0);
-    setDifficultyModalOpen(true);
-  };
-  const checkNumber = (pieceNum, fixedPosition) => {
-    if (pieceNum > fixedPosition + 20 || pieceNum < fixedPosition - 20) {
-      return false;
-    }
-    return true;
-  };
-  const handleDragStart = (e) => {
-    e.target.setAttrs({
-      shadowOffset: {
-        x: 15,
-        y: 15,
-      },
-      scaleX: 1.1,
-      scaleY: 1.1,
-    });
-    const { id } = e.target.attrs;
-    const items = shuffledPieceInfo.slice();
-    const item = items.find((i) => i.id === id);
-    const index = items.indexOf(item);
-    items.splice(index, 1);
-    items.push(item);
-    setShuffledPieceInfo(items);
-  };
-  const handleDragEnd = (e) => {
-    // console.log(e.target);
-    const update = {
-      duration: 0.5,
-      easing: Konva.Easings.ElasticEaseOut,
-      scaleX: 1,
-      scaleY: 1,
-      shadowOffsetX: 5,
-      shadowOffsetY: 5,
-    };
-    if (
-      checkNumber(e.target.attrs.x, e.target.attrs.cropX) &&
-      checkNumber(e.target.attrs.y, e.target.attrs.cropY)
-    ) {
-      setMatchPieceCount((count) => count + 1);
-      Object.assign(update, {
-        draggable: false,
-        x: e.target.attrs.cropX,
-        y: e.target.attrs.cropY,
-      });
-    }
-    e.target.to(update);
-  };
-
-  const [image] = useImage(`${process.env.PUBLIC_URL}/photo0000_6339.png`);
-
-  return (
-    <div>
-      <PuzzleGuide
-        matchPieceCount={matchPieceCount}
-        pieceTotalCount={pieceXCount * pieceYCount}
-        hour={hour}
-        minutes={minutes}
-        seconds={seconds}
-        handlePauseAction={handlePauseAction}
-      />
-      <Divider />
-      <PuzzleCanvas
-        image={image}
-        shuffledPieceInfo={shuffledPieceInfo}
-        handleDragStart={handleDragStart}
-        handleDragEnd={handleDragEnd}
-      />
-      <SelectDifficultyModal
-        open={difficultyModalOpen}
-        handleSelectDifficultyAction={handleSelectDifficultyAction}
-      />
-      <PauseModal
-        open={pauseModalOpen}
-        handlePauseReleseAction={handlePauseReleseAction}
-      />
-      <CompleteModal
-        open={completeModalOpen}
-        hour={hour}
-        minutes={minutes}
-        seconds={seconds}
-        handleRestartAction={handleRestartAction}
-      />
-    </div>
-  );
+  return <PuzzleTemplate imageUrl={imageUrl} />;
 };
 
 export default Puzzle;
